Fix SaveHeaderError.message attribute / drop `with_traceback`s:
Mopy/bash/bosh/__init__.py:
This is accessing a FileError attribute however SaveHeaderError inherits
from Exception - we might want to modify this
Anyway I replaced with_traceback with `from e` and did some more fstrings
around SaveHeaderError.
OK this `from` makes probably the traceback longer but should be more
informative - note this is in converters code which is a feature on hold
for now.