[!]Run wx.ArtProvider.GetBitmap before locale fidling RRR
!:

FromDIP was not called previously (in staticBitmap)

    bass.wx_bitmap['ART_WARNING'] = wxart(_wx.ART_WARNING,
        _wx.ART_MESSAGE_BOX, size=fromdip((32, 32), None))
    for size_tup in (siz16, (24, 24), (32, 32)):
        bass.wx_bitmap[('ART_UNDO', size_tup)] = wxart(_wx.ART_UNDO,
            _wx.ART_TOOLBAR, size=fromdip(size_tup, None))

- maybe call set_c_locale before the wx.ArtProvider - seems not needed

That's also the first stone of an image API - initially I used the wx ids
but hey, let's get started.
8 wx uses left in basher - it's mostly wx.Notebook
Note staticBitmap was returning the ART_UNDO while delegating to wx for
ART_WARN: net result staticBitmap almost gone :)
I am not sure that using the static FromDIP is correct. In that case we
should intercept accesses to wx_bitmaps dict and rescale the bitmap
(hoping that no other locale call will blow, aka that it was
_wx.ArtProvider.GetBitmap that blew only)

Under # 610, # 190 # 366 RRR