TTT fixup for BP dialog not being destroyed TTT
Well this is weird - on mac when clicking the X button on a modal dialog
Close evt is fired ok and on_closing handler is triggered - then the UI
hangs. Clicking on the UI makes the mac-error-sound and on the debugger
the _post method of the activate event is shown in the frames window,
*below the gui.top_level_windows.DialogWindow.show_modal* line:

        return self._native_widget.ShowModal() in (_wx.ID_OK, _wx.ID_YES)

So it seems the __exit__ is never called for those and we eternally stay
in the with ... as dialog statement.

Probably this is the issue with windows user who need to escape/enter the
invisible BP build dialog:
https://discord.com/channels/537706885965676554/537710082755133460/913812302975873066

Also (some of it is outdated but still):

https://stackoverflow.com/a/21391171/281545
https://stackoverflow.com/a/34300474/281545
https://stackoverflow.com/a/19815382/281545

We should encapsulate show_modal inside display_dialog

squash!     TTT fixup for BP dialog not being destroyed TTT
Mopy/bash/basher/mod_links.py:
SSS TTT master errors show_modal