Update the Mold messagge: WIP
Funny think I noted this in # 300 - we do need a beta soon

https://discord.com/channels/537706885965676554/537710082755133460/917717248850923530

Traceback (most recent call last):
  File "bash\bash.pyo", line 249, in main
  File "bash\bash.pyo", line 412, in _main
  File "bash\basher__init.pyo", line 4240, in Init
  File "bash\basher__init.pyo", line 4275, in InitData
  File "bash\bosh__init.pyo", line 1852, in init
  File "bash\load_order.pyo", line 93, in initialize_load_order_handle
  File "bash\load_order.pyo", line 327, in load_pickled_load_orders
  File "bash\bolt.pyo", line 1462, in load
Mold: Your settings in C:\Users\Preston\Documents\My Games\Skyrim\BashLoadOrders.dat come from an ancient Bash version. Please load them in 306 so they are converted to the newer format

ran wyre bash v308 and it gave me this error, i uninstalled it and installed the older 306 version like it says and ran it and the 306 ran flawlessly so i thought it was fixed so i uninstalled v306 and reinstalled v308 and again with the same error. can anyone help me?
nvm i just deleted the .dat file and it works now

WIP in the sense that we need still validate that the message says the
thruth - we don't want users to download ancient versions and still crash