Use a specialized subclass for "unknown_is_unused": SSS
This removes all the processing from the main Flags class and aligns
better with IntFlag, however I am not sure that Flags is-a IntFlag so
this is shelved for now.

Took the opportunity to add slots to subclasses missing it.
Added copy just in case - we should investigate if the memo dict is used
correctly in __deepcopy__