Correctly show ESL state for missing masters
Or attempt to show it as correctly as possible, by checking whether the
master was an ESL at the time the save was created.

No need to worry about checking the extension, the game will have
forced all .esl files to load in the ESL block back when the save was
created.