Add search bar and dropdown to doc browser RRR
Significantly improves the usability of the doc browser by letting you
select any of your installed plugins via a dropdown (and letting you
search through those plugins via a search bar that narrows down the
contents of the dropdown).

These two are also integrated with the rest of the doc browser - e.g. if
you select via the list on the left side, the search is cleared only if
necessary.

Closes # 628 <--- RRR