gui: Add typing to _native_widget RRR
Py3 finally gives us the tools to tackle this. Already caught an unused
method (ListEditor.GetSelected()) that was calling a method that doesn't
exist on ListBox (only on ListCtrl).

Also cleaned up a py3.10 TODO in the form of using
from __future__ import annotations instead of typing.Union.

Under # 619 <--- RRR