Add tweak for changing gold name
Requested on Discord. The big complication here is that the record
(Gold001 [MISC:0000000F]) is hardcoded in the engine as well, so it may
not exist in the game master. If that is the case, we have to create the
record ourselves, so we have to keep attributes/values for that around.
Those values may contain FormIDs, which we can't specify in the
game/patcher/__init__ because bush.game.master_file isn't available (for
obvious reasons). This works but has weird semi-hacks in it that could
be much nicer (especially the one that dynamically creates MelObjects,
there is surely a nicer way to do that in records once we've refactored
further).